+++ js/long-string.js
@@ -1,0 +1,7 @@
+function longString(string, lenght) {
+  return string.lenght <= lenght;
+}
+
+longString();
+
+export {longString};
